+++ frontend/pages/teachers.vue
@@ -43,10 +43,9 @@
                 <td>
                   <div class="flex gap-3 items-center py-3">
                     <button @click="$router.push(`/teachers/${teacher.id}/update`)" class="fab-mini-button bg-orange-500 text-white"><i class="bi bi-pencil-fill"></i></button>
-                    <button @click="showConfirmDelation = true" class="fab-mini-button bg-red-500 text-white"> <i class="bi bi-trash3-fill"></i></button>
+                    <button @click="deleteById(teacher.id)" class="fab-mini-button bg-red-500 text-white"> <i class="bi bi-trash3-fill"></i></button>
                   </div>
                 </td>
-                <ConfirmDeletion v-if="showConfirmDelation" @close="showConfirmDelation = false" @confirm="deleteById(teacher.id)"/>
               </tr>
               <tr v-if="!data.length">
                 <td colspan="5">
@@ -99,6 +98,7 @@
         this.data = dataReq;
       },
       async deleteById(id: number) {
+        if (!window.confirm('Realmente deseja deletar?')) return;
         await apiRequest(ApiConfig.baseUrlWith('users/'+id), null, undefined, undefined, {
           method: 'delete',
         });
